Use 64-bit integer type for particle number.

+++ src/options.hpp
@@ -51,7 +51,7 @@ public:
 
     void parse(int argc, char** argv);
 
-    unsigned int npart() const
+    uint64_t npart() const
     {
 	return npart_;
     }
@@ -118,7 +118,7 @@ public:
 
 private:
     /** number of particles */
-    unsigned int npart_;
+    uint64_t npart_;
     /** density */
     double density_;
     /** simulation timestep */
